Make Tile immutable
Mutable types are heap allocated; immutable types are better for
performance and should generally be preferred when the data in the
object doesn't need to change. In this case, `Tile`s are not mutated in
any way, so we'd might as well declare the type to be immutable.

Note also the change in how the game map is defined. When using mutable
objects, `fill` constructs an array containing all of the same object
but the comprehension ensures each item is a separate object. When using
immutable objects, the distinction doesn't matter (and doesn't really
make sense).

See https://docs.julialang.org/en/v1/manual/types/#Mutable-Composite-Types-1